Some minor bugfixes

+++ BFME2/WinFormsMainGUI.cs
@@ -261,6 +261,8 @@ namespace PatchLauncher
             catch (Exception ex)
             {
                 LogHelper.LoggerBFME2GUI.Error(ex.ToString());
+                MessageBox.Show("Something went wrong. Please see Logfiles for further Details. \n We will no reset the launcher state so you can close it. \n Please click on the discord Logo to get support.");
+                await TurnPatchesAndModsViewOn();
             }
 
             await TurnPatchesAndModsViewOn();
@@ -357,6 +359,8 @@ namespace PatchLauncher
                 catch (Exception ex)
                 {
                     LogHelper.LoggerBFME2GUI.Error(ex.ToString());
+                    MessageBox.Show("Something went wrong. Please see Logfiles for further Details. \n We will no reset the launcher state so you can close it. \n Please click on the discord Logo to get support.");
+                    await TurnPatchesAndModsViewOn();
                 }
             }
 
@@ -387,12 +391,12 @@ namespace PatchLauncher
 
                         languagePackSettings = JSONDataListHelper._DictionarylanguageSettings[Settings.Default.InstalledLanguageISOCode];
 
-                        RegistryService.WriteRegKeysInstallationBFME2(Settings.Default.GameInstallPath, languagePackSettings.RegistrySelectedLocale, languagePackSettings.RegistrySelectedLanguageName, languagePackSettings.RegistrySelectedLanguage);
+                        RegistryService.WriteRegKeysInstallationBFME25(Settings.Default.GameInstallPath, languagePackSettings.RegistrySelectedLocale, languagePackSettings.RegistrySelectedLanguageName, languagePackSettings.RegistrySelectedLanguage);
 
                         await InstallUpdatRepairRoutine(mainPack.FileName, mainPack.URLs, mainPack.MD5);
                         await InstallUpdatRepairRoutine(languagePackSettings.LanguagePackName, languagePackSettings.URLs, languagePackSettings.MD5);
 
-                        Settings.Default.PatchVersionInstalled = mainPack.LatestPatchVersionOfficial;
+                        Settings.Default.PatchVersionInstalled = mainPack.LatestPatchVersionOfficial; //patchPack.Version;
                         Settings.Default.IsGameInstalled = true;
                         Settings.Default.Save();
 
@@ -407,10 +411,6 @@ namespace PatchLauncher
                         }
 
                         taskPrepareInstallFolder.Dispose();
-
-                        SettingsToolStripMenuItem.Enabled = true;
-                        BFME25ToolStripMenuItem.Enabled = true;
-                        await TurnPatchesAndModsViewOn();
                     }
                 }
                 else
@@ -420,20 +420,18 @@ namespace PatchLauncher
             }
             catch (Exception ex)
             {
-                LogHelper.LoggerBFME2GUI.Error(ex.ToString());
+                LogHelper.LoggerBFME25GUI.Error(ex.ToString());
 
                 Settings.Default.IsGameInstalled = false;
                 BtnInstall.Text = Strings.BtnInstall_TextInstall;
                 Settings.Default.Save();
-
-                PibLoadingBorder.Visible = false;
-                PibLoadingRing.Visible = false;
-                LabelLoadingPanel.Visible = false;
-                IsLauncherCurrentlyWorking = false;
-                BtnInstall.Enabled = true;
-
-                Update();
             }
+
+            SettingsToolStripMenuItem.Enabled = true;
+            BFME25ToolStripMenuItem.Enabled = true;
+            await TurnPatchesAndModsViewOn();
+
+            Update();
         }
 
         private void BtnInstall_MouseLeave(object sender, EventArgs e)
@@ -573,46 +571,51 @@ namespace PatchLauncher
 
         private async Task InstallUpdatRepairRoutine(string ZIPFileName, string[] DownloadUrls, string CorrectMD5HashValue, bool hasExternalInstaller = false, bool externalInstallerHasLaunchAbility = false)
         {
-            PBarActualFile.Visible = true;
-            LblWorkerFileName.Visible = true;
-            LblWorkerIOTask.Visible = true;
-
-            Update();
-
-            PBarActualFile.Value = 0;
-            PBarActualFile.Maximum = 100;
-
-            string fullPathToZIPFile = Path.Combine(Application.StartupPath, ConstStrings.C_DOWNLOADFOLDER_NAME_BFME2, ZIPFileName);
-            string getFileNameFromArchiveIfExe = ConstStrings.C_BFME2_MAIN_GAME_FILE;
-
-            var progressHandlerDownload = new Progress<ProgressHelper>(progress =>
-            {
-                PBarActualFile.Value = (int)progress.PercentageValue;
-                LblWorkerFileName.Text = Path.GetFileNameWithoutExtension(ZIPFileName);
-                LblWorkerIOTask.Text = string.Concat(progress.ProgressedDownloadSizeInBytes / 1024000, " / ", progress.TotalDownloadSizeInBytes / 1024000, " MB @ ", Math.Round(progress.DownloadSpeedSizeInBytes / 1024000), " MB/s");
-            });
-
-            var progressHandlerExtraction = new Progress<ProgressHelper>(progress =>
-            {
-                PBarActualFile.Value = progress.CurrentlyExtractedFileCount;
-                PBarActualFile.Maximum = progress.TotalArchiveFileCount;
-                LblWorkerFileName.Text = progress.CurrentFileName;
-                LblWorkerIOTask.Text = string.Concat(progress.CurrentlyExtractedFileCount, " / ", progress.TotalArchiveFileCount);
-
-                if (hasExternalInstaller && Path.GetExtension(progress.CurrentFileName) == ".exe")
-                {
-                    getFileNameFromArchiveIfExe = progress.CurrentFileName!;
-                }
-            });
-
             try
             {
+                PBarActualFile.Visible = true;
+                LblWorkerFileName.Visible = true;
+                LblWorkerIOTask.Visible = true;
+
+                PBarActualFile.Value = 0;
+                PBarActualFile.Maximum = 100;
+
+                Update();
+
+                string fullPathToZIPFile = Path.Combine(Application.StartupPath, ConstStrings.C_DOWNLOADFOLDER_NAME_BFME2, ZIPFileName);
+                string getFileNameFromArchiveIfExe = ConstStrings.C_BFME2_MAIN_GAME_FILE;
+
+                var progressHandlerDownload = new Progress<ProgressHelper>(progress =>
+                {
+                    PBarActualFile.Value = (int)progress.PercentageValue;
+                    LblWorkerFileName.Text = Path.GetFileNameWithoutExtension(ZIPFileName);
+                    LblWorkerIOTask.Text = string.Concat(progress.ProgressedDownloadSizeInBytes / 1024000, " / ", progress.TotalDownloadSizeInBytes / 1024000, " MB @ ", Math.Round(progress.DownloadSpeedSizeInBytes / 1024000), " MB/s");
+                });
+
+                var progressHandlerExtraction = new Progress<ProgressHelper>(progress =>
+                {
+                    PBarActualFile.Value = progress.CurrentlyExtractedFileCount;
+                    PBarActualFile.Maximum = progress.TotalArchiveFileCount;
+                    LblWorkerFileName.Text = progress.CurrentFileName;
+                    LblWorkerIOTask.Text = string.Concat(progress.CurrentlyExtractedFileCount, " / ", progress.TotalArchiveFileCount);
+
+                    if (hasExternalInstaller && Path.GetExtension(progress.CurrentFileName) == ".exe")
+                    {
+                        getFileNameFromArchiveIfExe = progress.CurrentFileName!;
+                    }
+                });
+
                 BtnInstall.Text = Strings.BtnInstall_TextLaunch;
+
                 GameFileTools gameFileTools = new();
+
                 await gameFileTools.DownloadFile(Path.Combine(Application.StartupPath, ConstStrings.C_DOWNLOADFOLDER_NAME_BFME2), ZIPFileName, DownloadUrls, progressHandlerDownload);
+
                 LblWorkerFileName.Text = "";
                 LblWorkerIOTask.Text = "";
+
                 Update();
+
                 string calculatedMD5Value = MD5Tools.CalculateMD5(fullPathToZIPFile);
 
                 if (calculatedMD5Value == CorrectMD5HashValue)
@@ -621,10 +624,18 @@ namespace PatchLauncher
                 }
                 else
                 {
-                    LogHelper.LoggerBFME2GUI.Error(string.Format("MD5 HashSum check failed. Should be: {0} was: {1}", CorrectMD5HashValue, calculatedMD5Value));
+                    LogHelper.LoggerBFME2GUI.Error(string.Format("MD5 HashSum check failed. Should be: > {0} < was: > {1} <", CorrectMD5HashValue, calculatedMD5Value));
                     LogHelper.LoggerBFME2GUI.Information(string.Format("Deleting file > {0} < and retry Download...", ZIPFileName));
-                    File.Delete(fullPathToZIPFile);
+
+                    if (File.Exists(fullPathToZIPFile))
+                    {
+                        LogHelper.LoggerBFME2GUI.Information("File > {0} < existed and deleted!", fullPathToZIPFile);
+                        File.Delete(fullPathToZIPFile);
+                    }
+
+                    LogHelper.LoggerBFME2GUI.Information("Start downloading file: > {0} <", ZIPFileName);
                     await gameFileTools.DownloadFile(Path.Combine(Application.StartupPath, ConstStrings.C_DOWNLOADFOLDER_NAME_BFME2), ZIPFileName, DownloadUrls, progressHandlerDownload);
+
                     LogHelper.LoggerBFME2GUI.Information(string.Format("Now trying to extract > {0} <", ZIPFileName));
                     await gameFileTools.ExtractFile(Path.Combine(Application.StartupPath, ConstStrings.C_DOWNLOADFOLDER_NAME_BFME2), ZIPFileName, Settings.Default.GameInstallPath, progressHandlerExtraction, hasExternalInstaller);
                 }
@@ -643,6 +654,8 @@ namespace PatchLauncher
             catch (Exception ex)
             {
                 LogHelper.LoggerBFME2GUI.Error(ex.ToString());
+                MessageBox.Show("Something went wrong. Please see Logfiles for further Details. \n We will no reset the launcher state so you can close it. \n Please click on the discord Logo to get support.");
+                await TurnPatchesAndModsViewOn();
             }
         }
 
@@ -939,30 +952,39 @@ namespace PatchLauncher
 
         private async void RepairGameToolStripMenuItem_Click(object sender, EventArgs e)
         {
-            TurnPatchesAndModsViewOff();
-
-            if (Settings.Default.PatchVersionInstalled == 109)
+            try
             {
-                PatchPacks patchPacks = JSONDataListHelper._DictionaryPatchPacksSettings[Settings.Default.PatchVersionInstalled];
-                await PatchModDetectionHelper.MovePatch109FilesForBFME2(AssemblyNameHelper.BFMELauncherGameName, Path.GetFileNameWithoutExtension(patchPacks.FileName), true);
+                TurnPatchesAndModsViewOff();
+
+                if (Settings.Default.PatchVersionInstalled == 109)
+                {
+                    PatchPacks patchPacks = JSONDataListHelper._DictionaryPatchPacksSettings[Settings.Default.PatchVersionInstalled];
+                    await PatchModDetectionHelper.MovePatch109FilesForBFME2(AssemblyNameHelper.BFMELauncherGameName, Path.GetFileNameWithoutExtension(patchPacks.FileName), true);
+                }
+
+                LogHelper.LoggerGRepairFile.Information("Started Repairing...");
+                await RepairFileHelper.RepairFeature(AssemblyNameHelper.BFMELauncherGameName);
+
+                LogHelper.LoggerGRepairFile.Information("Downloading and/or extracting MainGame-Files if needed...");
+                await InstallUpdatRepairRoutine(mainPack.FileName, mainPack.URLs, mainPack.MD5);
+
+                LogHelper.LoggerGRepairFile.Information("Downloading and/or extracting Language-Files if needed...");
+                await InstallUpdatRepairRoutine(languagePackSettings.LanguagePackName, languagePackSettings.URLs, languagePackSettings.MD5);
+
+                Settings.Default.PatchVersionInstalled = mainPack.LatestPatchVersionOfficial;
+                Settings.Default.ActivePatchOrModExternalProgramFolderPath = "";
+                Settings.Default.Save();
+
+                UpdatePanelButtonActiveState();
+
+                await TurnPatchesAndModsViewOn();
             }
-
-            LogHelper.LoggerGRepairFile.Information("Started Repairing...");
-            await RepairFileHelper.RepairFeature(AssemblyNameHelper.BFMELauncherGameName);
-
-            LogHelper.LoggerGRepairFile.Information("Downloading and/or extracting MainGame-Files if needed...");
-            await InstallUpdatRepairRoutine(mainPack.FileName, mainPack.URLs, mainPack.MD5);
-
-            LogHelper.LoggerGRepairFile.Information("Downloading and/or extracting Language-Files if needed...");
-            await InstallUpdatRepairRoutine(languagePackSettings.LanguagePackName, languagePackSettings.URLs, languagePackSettings.MD5);
-
-            Settings.Default.PatchVersionInstalled = mainPack.LatestPatchVersionOfficial;
-            Settings.Default.ActivePatchOrModExternalProgramFolderPath = "";
-            Settings.Default.Save();
-
-            UpdatePanelButtonActiveState();
-
-            await TurnPatchesAndModsViewOn();
+            catch (Exception ex)
+            {
+                LogHelper.LoggerGRepairFile.Error(ex.ToString());
+                MessageBox.Show("Something went wrong. Please see Logfiles for further Details. \n We will no reset the launcher state so you can close it. \n Please click on the discord Logo to get support.");
+                await TurnPatchesAndModsViewOn();
+            }
         }
 
         private void MenuItemLaunchGame_Click(object sender, EventArgs e)
